♻️ refactor: 폴더 구조 전체적으로 수정
