+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router } from "react-router-dom"
 import Header from "./widgets/common/ui/Header.tsx"
 import Footer from "./widgets/common/ui/Footer.tsx"
-import PostsManagerPage from "./pages/PostsManagerPage.tsx"
+import PostsManagerPage from "@pages/PostsManagerPage"
 
 const App = () => {
   return (
