--- src/entities/common/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./LoadingSpinner"
