--- src/features/comment/comment-list/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./CommentsList"
