+++ src/features/comment/ui/comment-form/CommentForm.tsx
@@ -1,6 +1,5 @@
-import { Textarea } from "@shared/ui"
+import { Textarea, Button } from "@shared/ui"
 import { Comment } from "@entities/comment/model"
-import { DialogActions } from "@features/common/dialog/ui"
 
 interface CommentFormProps {
   mode: "add" | "edit"
@@ -9,12 +8,7 @@ interface CommentFormProps {
   onSubmit: () => void
 }
 
-export const CommentForm = ({
-  mode,
-  comment,
-  onBodyChange,
-  onSubmit,
-}: CommentFormProps) => {
+export const CommentForm = ({ mode, comment, onBodyChange, onSubmit }: CommentFormProps) => {
   const isEdit = mode === "edit"
 
   return (
@@ -24,7 +18,9 @@ export const CommentForm = ({
         value={isEdit ? (comment?.body ?? "") : ""}
         onChange={(e) => onBodyChange(e.target.value)}
       />
-      <DialogActions onSubmit={onSubmit} submitText={isEdit ? "댓글 업데이트" : "댓글 추가"} />
+      <div className="flex justify-end gap-2 mt-4">
+        <Button onClick={onSubmit}>{isEdit ? "댓글 업데이트" : "댓글 추가"}</Button>
+      </div>
     </div>
   )
 }
