+++ src/features/comment/ui/comment-list/index.ts
@@ -1,1 +1,2 @@
+export * from "./CommentsList"
 export * from "./PostCommentsFeature"
