+++ src/features/comment/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./comment-form"
+export * from "./comment-list"
+export * from "./comment-management"
