--- src/features/common/dialog/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./DialogActions"
