--- src/features/common/pagination/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./PaginationControls"
