--- src/features/post/post-management/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./PostActionButtons"
