+++ src/features/post/ui/post-search/SearchInput.tsx
@@ -1,4 +1,4 @@
-import { Input } from "@shared/ui/Input"
+import { Input } from "@shared/ui"
 import { Search } from "lucide-react"
 
 interface SearchInputProps {
