+++ src/features/user/ui/user-info/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserInfo"
