--- src/features/user/user-info/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./UserInfo"
