+++ src/shared/ui/Dialog.tsx
@@ -1,4 +1,3 @@
-import * as React from "react"
 import { forwardRef } from "react"
 import * as DialogPrimitive from "@radix-ui/react-dialog"
 import { X } from "lucide-react"
