+++ src/shared/ui/Table.tsx
@@ -1,4 +1,3 @@
-import * as React from "react"
 import { forwardRef } from "react"
 
 interface TableProps extends React.TableHTMLAttributes<HTMLTableElement> {
