+++ src/shared/ui/Textarea.tsx
@@ -1,4 +1,3 @@
-import * as React from "react"
 import { forwardRef } from "react"
 
 interface TextareaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
