+++ src/shared/ui/index.ts
@@ -5,3 +5,6 @@ export { Card } from './Card'
 export { Select } from './Select'
 export { Dialog } from './Dialog'
 export { Table } from './Table'
+
+export { LoadingSpinner } from './LoadingSpinner'
+export { Pagination } from './Pagination'
