+++ src/widgets/common/ui/index.ts
@@ -1,2 +1,2 @@
-export * from "./Header"
-export * from "./Footer"
+export { default as Header } from "./Header"
+export { default as Footer } from "./Footer"
